Move snippet entity
