+++ front/src/app/snippets/snippet.ts
